# perf(cache): Add a sharded map for global cache