--- a/go.mod
+++ b/go.mod
@@ -2,6 +2,8 @@ module github.com/dgraph-io/dgraph/v24
 
 go 1.22.7
 
+replace github.com/dgraph-io/ristretto => /home/harshil/Projects/ristretto/
+
 require (
 	contrib.go.opencensus.io/exporter/jaeger v0.2.1
 	contrib.go.opencensus.io/exporter/prometheus v0.4.2
@@ -15,6 +17,7 @@ require (
 	github.com/dgraph-io/gqlgen v0.13.2
 	github.com/dgraph-io/gqlparser/v2 v2.2.2
 	github.com/dgraph-io/graphql-transport-ws v0.0.0-20210511143556-2cef522f1f15
+	github.com/dgraph-io/ristretto v0.0.0-00010101000000-000000000000
 	github.com/dgraph-io/ristretto/v2 v2.0.0
 	github.com/dgraph-io/simdjson-go v0.3.0
 	github.com/dgryski/go-farm v0.0.0-20240924180020-3414d57e47da