--- a/posting/mvcc.go
+++ b/posting/mvcc.go
@@ -19,8 +19,10 @@ package posting
 import (
 	"bytes"
 	"encoding/hex"
+	"fmt"
+	"math"
+	"runtime"
 	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -34,6 +36,7 @@ import (
 	"github.com/dgraph-io/dgo/v240/protos/api"
 	"github.com/dgraph-io/dgraph/v24/protos/pb"
 	"github.com/dgraph-io/dgraph/v24/x"
+	"github.com/dgraph-io/ristretto"
 	"github.com/dgraph-io/ristretto/v2/z"
 )
 
@@ -63,12 +66,7 @@ type CachePL struct {
 	count      int
 	list       *List
 	lastUpdate uint64
-}
-
-type GlobalCache struct {
-	sync.RWMutex
-
-	items map[string]*CachePL
+	lastRead   time.Time
 }
 
 var (
@@ -85,10 +83,12 @@ var (
 		priorityKeys: make([]*pooledKeys, 2),
 	}
 
-	globalCache = &GlobalCache{items: make(map[string]*CachePL, 100)}
+	memoryLayer = initMemoryLayer()
+	numShards   = 256
 )
 
 func init() {
+	runtime.SetCPUProfileRate(200)
 	x.AssertTrue(len(IncrRollup.priorityKeys) == 2)
 	for i := range IncrRollup.priorityKeys {
 		IncrRollup.priorityKeys[i] = &pooledKeys{
@@ -134,13 +134,9 @@ func (ir *incrRollupi) rollUpKey(writer *TxnWriter, key []byte) error {
 	}
 
 	RemoveCacheFor(key)
-
-	globalCache.Lock()
-	val, ok := globalCache.items[string(key)]
-	if ok {
-		val.list = nil
-	}
-	globalCache.Unlock()
+	//pk, _ := x.Parse(key)
+	//fmt.Println("====Setting cache delete rollup", ts, pk)
+	memoryLayer.Del(z.MemHash(key))
 	// TODO Update cache with rolled up results
 	// If we do a rollup, we typically won't need to update the key in cache.
 	// The only caveat is that the key written by rollup would be written at +1
@@ -175,6 +171,26 @@ func (ir *incrRollupi) addKeyToBatch(key []byte, priority int) {
 	}
 }
 
+func (ir *incrRollupi) mem(getNewTs func(bool) uint64) {
+	forceRollupTick := time.NewTicker(500 * time.Millisecond)
+	defer forceRollupTick.Stop()
+	deleteCacheTick := time.NewTicker(1 * time.Second)
+	defer deleteCacheTick.Stop()
+
+	for {
+		select {
+		case <-deleteCacheTick.C:
+			memoryLayer.deleteOldItems(ir.getNewTs(false))
+		case <-forceRollupTick.C:
+			t := 10000
+			memoryLayer.insert += t
+			if memoryLayer.insert > 5*t {
+				memoryLayer.insert = 5 * t
+			}
+		}
+	}
+}
+
 // Process will rollup batches of 64 keys in a go routine.
 func (ir *incrRollupi) Process(closer *z.Closer, getNewTs func(bool) uint64) {
 	ir.getNewTs = getNewTs
@@ -192,6 +208,8 @@ func (ir *incrRollupi) Process(closer *z.Closer, getNewTs func(bool) uint64) {
 	defer cleanupTick.Stop()
 	forceRollupTick := time.NewTicker(500 * time.Millisecond)
 	defer forceRollupTick.Stop()
+
+	ir.mem(getNewTs)
 
 	doRollup := func(batch *[][]byte, priority int) {
 		currTs := time.Now().Unix()
@@ -318,6 +336,7 @@ func (txn *Txn) CommitToDisk(writer *TxnWriter, commitTs uint64) error {
 			for ; idx < len(keys); idx++ {
 				key := keys[idx]
 				data := cache.deltas[key]
+				//fmt.Println("--------------------------------------------------------HUI")
 				if len(data) == 0 {
 					continue
 				}
@@ -349,9 +368,7 @@ func ResetCache() {
 	if lCache != nil {
 		lCache.Clear()
 	}
-	globalCache.Lock()
-	globalCache.items = make(map[string]*CachePL)
-	globalCache.Unlock()
+	memoryLayer.Clear()
 }
 
 // RemoveCacheFor will delete the list corresponding to the given key.
@@ -362,11 +379,200 @@ func RemoveCacheFor(key []byte) {
 	}
 }
 
+type setItems struct {
+	keyHash uint64
+	list    *List
+	readTs  uint64
+}
+
+type MemoryLayer struct {
+	shards []*lockedMap
+	setBuf chan *setItems
+	skull  *ristretto.Skull[uint64, struct{}]
+
+	insert            int
+	numCacheRead      int
+	numCacheReadFails int
+	numDisksRead      int
+	numCacheSave      int
+}
+
+func initMemoryLayer() *MemoryLayer {
+	sm := &MemoryLayer{
+		shards: make([]*lockedMap, numShards),
+		setBuf: make(chan *setItems, 32*1024),
+		skull:  ristretto.GetSkull[uint64, struct{}](),
+	}
+	for i := range sm.shards {
+		sm.shards[i] = newLockedMap()
+	}
+	return sm
+}
+
+func (sm *MemoryLayer) get(key uint64) (*CachePL, bool) {
+	return sm.shards[key%uint64(numShards)].get(key)
+}
+
+func (sm *MemoryLayer) set(key uint64, i *CachePL) {
+	if i == nil {
+		// If item is nil make this Set a no-op.
+		return
+	}
+
+	sm.shards[key%uint64(numShards)].set(key, i)
+}
+
+func (sm *MemoryLayer) del(key uint64) {
+	sm.shards[key%uint64(numShards)].del(key)
+}
+
+func (sm *MemoryLayer) Get(key uint64) (*CachePL, bool) {
+	return sm.shards[key%uint64(numShards)].Get(key)
+}
+
+func (sm *MemoryLayer) Set(key uint64, i *CachePL) {
+	if i == nil {
+		// If item is nil make this Set a no-op.
+		return
+	}
+
+	sm.shards[key%uint64(numShards)].Set(key, i)
+}
+
+func (sm *MemoryLayer) Del(key uint64) {
+	sm.shards[key%uint64(numShards)].Del(key)
+	sm.skull.Del(key)
+}
+
+func (sm *MemoryLayer) UnlockKey(key uint64) {
+	sm.shards[key%uint64(numShards)].Unlock()
+}
+
+func (sm *MemoryLayer) LockKey(key uint64) {
+	sm.shards[key%uint64(numShards)].Lock()
+}
+
+func (sm *MemoryLayer) RLockKey(key uint64) {
+	sm.shards[key%uint64(numShards)].RLock()
+}
+
+func (sm *MemoryLayer) RUnlockKey(key uint64) {
+	sm.shards[key%uint64(numShards)].RUnlock()
+}
+
+func (sm *MemoryLayer) Clear() {
+	for i := 0; i < numShards; i++ {
+		sm.shards[i].Clear()
+	}
+}
+
+type lockedMap struct {
+	sync.RWMutex
+	data map[uint64]*CachePL
+}
+
+func newLockedMap() *lockedMap {
+	return &lockedMap{
+		data: make(map[uint64]*CachePL),
+	}
+}
+
+func (m *lockedMap) get(key uint64) (*CachePL, bool) {
+	item, ok := m.data[key]
+	return item, ok
+}
+
+func (m *lockedMap) Get(key uint64) (*CachePL, bool) {
+	m.RLock()
+	defer m.RUnlock()
+	item, ok := m.data[key]
+	return item, ok
+}
+
+func (m *lockedMap) set(key uint64, i *CachePL) {
+	if i == nil {
+		// If the item is nil make this Set a no-op.
+		return
+	}
+
+	m.data[key] = i
+}
+
+func (m *lockedMap) Set(key uint64, i *CachePL) {
+	m.Lock()
+	defer m.Unlock()
+	m.set(key, i)
+}
+
+func (m *lockedMap) del(key uint64) {
+	delete(m.data, key)
+}
+
+func (m *lockedMap) Del(key uint64) {
+	m.Lock()
+	if l, ok := m.data[key]; ok && l != nil {
+		l.list = nil
+	}
+	m.Unlock()
+}
+
+func (m *lockedMap) Clear() {
+	m.Lock()
+	m.data = make(map[uint64]*CachePL)
+	m.Unlock()
+}
+
 func NewCachePL() *CachePL {
 	return &CachePL{
 		count:      0,
 		list:       nil,
 		lastUpdate: 0,
+	}
+}
+
+func checkForRollup(key []byte, l *List) {
+	deltaCount := l.mutationMap.len()
+	// If deltaCount is high, send it to high priority channel instead.
+	if deltaCount > 500 {
+		IncrRollup.addKeyToBatch(key, 0)
+	}
+}
+
+func (ml *MemoryLayer) updateItemInCache(key string, pk x.ParsedKey, delta []byte, startTs, commitTs uint64) {
+	if commitTs == 0 {
+		return
+	}
+
+	p := new(pb.PostingList)
+	x.Check(proto.Unmarshal(delta, p))
+	//fmt.Println("======COMMITTING", startTs, commitTs, pk, p)
+
+	keyHash := z.MemHash([]byte(key))
+	// TODO under the same lock
+	ml.LockKey(keyHash)
+	defer ml.UnlockKey(keyHash)
+
+	a := 1
+	if a == 1 {
+		delete(ml.shards[keyHash%uint64(numShards)].data, keyHash)
+		return
+	}
+
+	val, ok := ml.get(keyHash)
+	if !ok {
+		val = NewCachePL()
+		val.lastUpdate = commitTs
+		ml.set(keyHash, val)
+		return
+	}
+
+	val.lastUpdate = commitTs
+	val.count -= 1
+
+	if val.list != nil {
+		val.list.setMutationAfterCommit(startTs, commitTs, p, true)
+		checkForRollup([]byte(key), val.list)
+		//fmt.Println("====Setting cache list", commitTs, pk, p, val.list.mutationMap, val.list.key)
 	}
 }
 
@@ -378,34 +584,12 @@ func (txn *Txn) UpdateCachedKeys(commitTs uint64) {
 
 	for key, delta := range txn.cache.deltas {
 		RemoveCacheFor([]byte(key))
+
 		pk, _ := x.Parse([]byte(key))
 		if !ShouldGoInCache(pk) {
 			continue
 		}
-		globalCache.Lock()
-		val, ok := globalCache.items[key]
-		if !ok {
-			val = NewCachePL()
-			val.lastUpdate = commitTs
-			globalCache.items[key] = val
-		}
-		if commitTs != 0 {
-			// TODO Delete this if the values are too old in an async thread
-			val.lastUpdate = commitTs
-		}
-		if !ok {
-			globalCache.Unlock()
-			continue
-		}
-
-		val.count -= 1
-
-		if commitTs != 0 && val.list != nil {
-			p := new(pb.PostingList)
-			x.Check(proto.Unmarshal(delta, p))
-			val.list.setMutationAfterCommit(txn.StartTs, commitTs, p, true)
-		}
-		globalCache.Unlock()
+		memoryLayer.updateItemInCache(key, pk, delta, txn.StartTs, commitTs)
 	}
 }
 
@@ -433,6 +617,7 @@ func ReadPostingList(key []byte, it *badger.Iterator) (*List, error) {
 	// lists ended up being rolled-up multiple times. This issue was caught by the
 	// uid-set Jepsen test.
 	pk, err := x.Parse(key)
+	//fmt.Println("READING KEY", key, pk)
 	if err != nil {
 		return nil, errors.Wrapf(err, "while reading posting list with key [%v]", key)
 	}
@@ -449,6 +634,7 @@ func ReadPostingList(key []byte, it *badger.Iterator) (*List, error) {
 	l := new(List)
 	l.key = key
 	l.plist = new(pb.PostingList)
+	l.minTs = 0
 
 	// We use the following block of code to trigger incremental rollup on this key.
 	deltaCount := 0
@@ -477,19 +663,21 @@ func ReadPostingList(key []byte, it *badger.Iterator) (*List, error) {
 
 		switch item.UserMeta() {
 		case BitEmptyPosting:
-			l.minTs = item.Version()
 			return l, nil
 		case BitCompletePosting:
 			if err := unmarshalOrCopy(l.plist, item); err != nil {
 				return nil, err
 			}
-			l.minTs = item.Version()
 
+			l.minTs = item.Version()
 			// No need to do Next here. The outer loop can take care of skipping
 			// more versions of the same key.
 			return l, nil
 		case BitDeltaPosting:
 			err := item.Value(func(val []byte) error {
+				if l.mutationMap == nil {
+					l.mutationMap = newMutableLayer()
+				}
 				pl := &pb.PostingList{}
 				if err := proto.Unmarshal(val, pl); err != nil {
 					return err
@@ -540,65 +728,133 @@ func (c *CachePL) Set(l *List, readTs uint64) {
 }
 
 func ShouldGoInCache(pk x.ParsedKey) bool {
-	return (!pk.IsData() && strings.HasSuffix(pk.Attr, "dgraph.type"))
+	//return !pk.IsData()
+	return true
+	//return false
 }
 
 func PostingListCacheEnabled() bool {
-	return lCache != nil
+	return false
+	//return lCache != nil
 }
 
-func GetNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
-	return getNew(key, pstore, readTs)
-}
-
-func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
-	if PostingListCacheEnabled() {
-		l, ok := lCache.Get(key)
-		if ok && l != nil {
-			// No need to clone the immutable layer or the key since mutations will not modify it.
-			lCopy := &List{
-				minTs: l.minTs,
-				maxTs: l.maxTs,
-				key:   key,
-				plist: l.plist,
-			}
-			l.RLock()
-			lCopy.mutationMap = l.mutationMap.clone()
-			l.RUnlock()
-			return lCopy, nil
-		}
+func (ml *MemoryLayer) Process(i *setItems) {
+	if ml.insert < 0 {
+		return
 	}
+	ml.insert -= 1
+	victims, add := ml.skull.Set(i.keyHash, int64(i.list.ApproxLen()/100))
+	//fmt.Println(add, i.list.ApproxLen())
 
-	if pstore.IsClosed() {
-		return nil, badger.ErrDBClosed
-	}
-
-	pk, _ := x.Parse(key)
-
-	if ShouldGoInCache(pk) {
-		globalCache.Lock()
-		cacheItem, ok := globalCache.items[string(key)]
+	if add {
+		ml.LockKey(i.keyHash)
+		i.list.RLock()
+		cacheItem, ok := ml.get(i.keyHash)
 		if !ok {
-			cacheItem = NewCachePL()
-			globalCache.items[string(key)] = cacheItem
-		}
-		cacheItem.count += 1
+			cacheItemNew := NewCachePL()
+			cacheItemNew.count = 1
+			cacheItemNew.list = copyList(i.list)
+			cacheItemNew.lastUpdate = i.list.maxTs
+			ml.set(i.keyHash, cacheItemNew)
+		} else {
+			// Only set l to the cache if readTs >= latestTs, which implies that l is
+			// the latest version of the PL. We also check that we're reading a version
+			// from Badger, which is higher than the write registered by the cache.
 
-		// We use badger subscription to invalidate the cache. For every write we make the value
-		// corresponding to the key in the cache to nil. So, if we get some non-nil value from the cache
-		// then it means that no  writes have happened after the last set of this key in the cache.
-		if ok {
-			if cacheItem.list != nil && cacheItem.list.minTs <= readTs {
-				cacheItem.list.RLock()
-				lCopy := copyList(cacheItem.list)
-				cacheItem.list.RUnlock()
-				globalCache.Unlock()
-				return lCopy, nil
-			}
+			//fmt.Println("====Setting cache", readTs, pk, l.mutationMap)
+			cacheItem.Set(copyList(i.list), i.readTs)
 		}
-		globalCache.Unlock()
+		ml.numCacheSave += 1
+		//allV, _ := i.list.AllValues(i.readTs)
+		//uids, _ := i.list.Uids(ListOptions{ReadTs: i.readTs})
+		//fmt.Println("====Setting into cache", i.readTs, i.list.key, i.list.mutationMap, allV, uids)
+		i.list.RUnlock()
+
+		//idx := int(i.keyHash % uint64(numShards))
+		//if len(ml.shards[idx].data) > 500 {
+		//	for keyHash, pl := range ml.shards[idx].data {
+		//		if pl.lastRead < i.readTs-100 {
+		//			delete(ml.shards[idx].data, keyHash)
+		//		}
+		//	}
+		//}
+		ml.UnlockKey(i.keyHash)
 	}
 
+	for _, vic := range victims {
+		ml.LockKey(vic.Key)
+		delete(ml.shards[vic.Key%uint64(numShards)].data, vic.Key)
+		ml.UnlockKey(vic.Key)
+	}
+}
+
+func (ml *MemoryLayer) deleteOldItems(ts uint64) {
+	fmt.Println("Deleting old items", ml.numCacheRead, ml.numDisksRead, ml.numCacheSave, ml.numCacheReadFails, float64(ml.numCacheRead)/float64(ml.numDisksRead), ml.insert)
+	lb := 0
+	la := 0
+	t1 := time.Now()
+	defer func() {
+		fmt.Println("Done deleting old items", lb, la, time.Since(t1))
+	}()
+
+	//ml.skull.ExpiryMap.SkullCleanup(func(keyHash uint64) {
+	//	ml.LockKey(keyHash)
+	//	delete(ml.shards[keyHash%uint64(numShards)].data, keyHash)
+	//	ml.UnlockKey(keyHash)
+	//}, ml.skull.CachePolicy)
+
+	for i := 0; i < numShards; i++ {
+		ml.shards[i].Lock()
+		lb += len(ml.shards[i].data)
+		for keyHash, pl := range ml.shards[i].data {
+			if time.Since(pl.lastRead) > 5*time.Second {
+				delete(ml.shards[i].data, keyHash)
+				ml.skull.Del(keyHash)
+			}
+			if len(ml.shards[i].data) < 500 { // Keeps like 200k entries after this
+				break
+			}
+		}
+		la += len(ml.shards[i].data)
+		ml.shards[i].Unlock()
+	}
+}
+
+func (ml *MemoryLayer) saveInCache(keyHash, readTs uint64, l *List) {
+	ml.Process(&setItems{
+		keyHash: keyHash,
+		readTs:  readTs,
+		list:    l,
+	})
+}
+
+func (ml *MemoryLayer) readFromCache(key []byte, keyHash, readTs uint64) *List {
+	ml.RLockKey(keyHash)
+
+	ml.skull.Get(keyHash)
+	cacheItem, ok := ml.get(keyHash)
+
+	if ok {
+		cacheItem.count += 1
+		cacheItem.lastRead = time.Now()
+		if cacheItem.list != nil && cacheItem.list.minTs <= readTs {
+			cacheItem.list.RLock()
+			lCopy := copyList(cacheItem.list)
+			cacheItem.list.RUnlock()
+			ml.RUnlockKey(keyHash)
+			checkForRollup(key, lCopy)
+			//allV, _ := lCopy.AllValues(readTs)
+			//uids, _ := lCopy.Uids(ListOptions{ReadTs: readTs})
+			//fmt.Println("====Getting cache", readTs, lCopy.key, lCopy.mutationMap, allV, uids)
+			return lCopy
+		}
+	}
+	ml.RUnlockKey(keyHash)
+	return nil
+}
+
+func (ml *MemoryLayer) readFromDisk(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
+	ml.numDisksRead += 1
 	txn := pstore.NewTransactionAt(readTs, false)
 	defer txn.Discard()
 
@@ -611,31 +867,86 @@ func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
 	defer itr.Close()
 	itr.Seek(key)
 	l, err := ReadPostingList(key, itr)
+	//fmt.Println("=============GETTING DISK", key, l.mutationMap, l.plist)
+	if err != nil {
+		return l, err
+	}
+	return l, nil
+}
+
+func (ml *MemoryLayer) ReadData(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
+	pk, _ := x.Parse(key)
+	var keyHash uint64
+
+	gic := ShouldGoInCache(pk)
+	if gic {
+		keyHash = z.MemHash(key)
+		l := ml.readFromCache(key, keyHash, readTs)
+		if l != nil {
+			//fmt.Println(pk, pk.IsData())
+			ml.numCacheRead += 1
+			return l, nil
+		} else {
+			ml.numCacheReadFails += 1
+		}
+		l, err := ml.readFromDisk(key, pstore, math.MaxUint64)
+		//fmt.Println("READING FROM DISK", l.minTs, readTs)
+		if err != nil {
+			return nil, err
+		}
+		ml.saveInCache(keyHash, readTs, l)
+		if l.minTs == 0 || readTs >= l.minTs {
+			return l, nil
+		}
+	}
+
+	l, err := ml.readFromDisk(key, pstore, readTs)
+	if err != nil {
+		return nil, err
+	}
+
+	return l, nil
+}
+
+func GetNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
+	return getNew(key, pstore, readTs)
+}
+
+func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {
+	//fmt.Println("Get new", key)
+	if PostingListCacheEnabled() {
+		l, ok := lCache.Get(key)
+		if ok && l != nil {
+			// No need to clone the immutable layer or the key since mutations will not modify it.
+			memoryLayer.numCacheRead += 1
+			lCopy := &List{
+				minTs: l.minTs,
+				maxTs: l.maxTs,
+				key:   key,
+				plist: l.plist,
+			}
+			l.RLock()
+			if l.mutationMap != nil {
+				lCopy.mutationMap = l.mutationMap.clone()
+			}
+			l.RUnlock()
+			return lCopy, nil
+		} else {
+			memoryLayer.numCacheReadFails += 1
+		}
+	}
+
+	if pstore.IsClosed() {
+		return nil, badger.ErrDBClosed
+	}
+
+	l, err := memoryLayer.ReadData(key, pstore, readTs)
 	if err != nil {
 		return l, err
 	}
 
-	// Only set l to the cache if readTs >= latestTs, which implies that l is
-	// the latest version of the PL. We also check that we're reading a version
-	// from Badger, which is higher than the write registered by the cache.
-	if ShouldGoInCache(pk) {
-		globalCache.Lock()
-		l.RLock()
-		cacheItem, ok := globalCache.items[string(key)]
-		if !ok {
-			cacheItemNew := NewCachePL()
-			cacheItemNew.count = 1
-			cacheItemNew.list = copyList(l)
-			cacheItemNew.lastUpdate = l.maxTs
-			globalCache.items[string(key)] = cacheItemNew
-		} else {
-			cacheItem.Set(copyList(l), readTs)
-		}
-		l.RUnlock()
-		globalCache.Unlock()
-	}
-
 	if PostingListCacheEnabled() {
+		memoryLayer.numCacheSave += 1
 		lCache.Set(key, l, 0)
 	}