--- a/worker/online_restore.go
+++ b/worker/online_restore.go
@@ -410,6 +410,7 @@ func handleRestoreProposal(ctx context.Context, req *pb.RestoreRequest, pidx uin
 		return errors.Wrapf(err, "cannot load schema after restore")
 	}
 
+	posting.ResetCache()
 	ResetAclCache()
 
 	// Reset gql schema only when the restore is not partial, so that after this restore