Main without risteretto cache:
```
badger 2024/10/06 20:34:31 INFO: All 0 tables opened in 0s
badger 2024/10/06 20:34:31 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/06 20:34:31 INFO: Set nextTxnTs to 0
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/v24/posting
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkTestCache-8     1673727               751.5 ns/op
PASS
ok      github.com/dgraph-io/dgraph/v24/posting 2.073s
```

Main with risteretto branch:
```
badger 2024/10/06 20:35:36 INFO: All 0 tables opened in 0s
badger 2024/10/06 20:35:36 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/06 20:35:36 INFO: Set nextTxnTs to 0
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/v24/posting
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkTestCache-8     2333268               501.3 ns/op
PASS
ok      github.com/dgraph-io/dgraph/v24/posting 1.531s
```


This PR:
```
badger 2024/10/06 20:34:12 INFO: All 0 tables opened in 0s
badger 2024/10/06 20:34:12 INFO: Discard stats nextEmptySlot: 0
badger 2024/10/06 20:34:12 INFO: Set nextTxnTs to 0
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/v24/posting
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkTestCache-8     3668409               319.3 ns/op
PASS
ok      github.com/dgraph-io/dgraph/v24/posting 1.576s
```